Created new component for right side section of card info page...
Also, updated lazy loading and card image loading timeouts to increase responsiveness
