+++ src/components/card/CardInformation.js
@@ -1,5 +1,5 @@
-import React, { useState, useEffect, lazy, Suspense, useMemo, memo } from 'react'
-import { Grid, Chip } from '@material-ui/core'
+import React, { useState, useEffect, lazy, Suspense, useMemo } from 'react'
+import { Chip } from '@material-ui/core'
 import {Helmet} from 'react-helmet'
 
 import { handleFetch } from '../../helper/FetchHandler'
@@ -11,11 +11,26 @@ import OneThirdTwoThirdsGrid from '../grid/OneThirdTwoThirdsGrid'
 
 import Breadcrumb from '../Breadcrumb'
 import CardData from './CardData'
-import {RightBoxPaper, RightBoxHeaderTypography, RightBoxSubHeaderTypography} from '../grid/OneThirdTwoThirdsGrid'
-import {DarkTranslucentDivider} from '../util/Divider'
 
-const CardInformationSection = lazy( () => import('./CardInformationSection') )
-const Footer = lazy( () => import('../Footer') )
+const CardInformationRelatedContent = lazy( () => import('./CardInformationRelatedContent') )
+
+const crumbs = ['Home', 'Card Browse']
+
+async function populateBanListChips(banListInfo)
+{
+	return banListInfo.map( (item, index) => <Chip key={index} label={`${item.banListDate}  •  ${item.banStatus.charAt(0)}`} />)
+}
+
+async function populateProductChips(productInfo, cardID)
+{
+	return productInfo.map( (product, index) => {
+		return product.productContent.map(item => <Chip
+				key={index}
+				label={`${product.productReleaseDate}  •  ${product.productId} #${item.productPosition}  •  ${item.rarities.join(', ')}`}
+				onClick={ () => setTimeout( () => window.location.assign(`/product/${product.productId}#${cardID}`), 250 ) }
+		/>)
+		})
+}
 
 
 const Card = ( { match, history } ) =>
@@ -40,7 +55,7 @@ const Card = ( { match, history } ) =>
 	const [productInfoChips, setPackInfoChips] = useState([])
 	const [banListInfoChips, setBanListInfoChips] = useState([])
 
-	const [dynamicCrumbs, setDynamicCrumbs] = useState(['Home', 'Card Browse', ''])
+	const [dynamicCrumbs, setDynamicCrumbs] = useState([...crumbs, ''])
 
 	const helmetData = useMemo( () => {
 		handleFetch(`${NAME_maps_ENDPOINT['cardInstanceUrl']}${cardID}?allInfo=true`, history, (json) => {
@@ -56,8 +71,7 @@ const Card = ( { match, history } ) =>
 			setPackInfo( (json.foundIn === undefined)? [] : json.foundIn )
 			setBanListInfo( (json.restrictedIn === undefined)? [] : json.restrictedIn )
 
-			const crumbs = ['Home', 'Card Browse', json.cardID]
-			setDynamicCrumbs(crumbs)
+			setDynamicCrumbs([...crumbs, json.cardID])
 			setIsLoading(false)
 		})
 
@@ -79,18 +93,7 @@ const Card = ( { match, history } ) =>
 	useEffect( () => {
 		if ( productInfoChips.length === 0 )
 		{
-			async function populateProductChips(productInfo)
-			{
-				return productInfo.map( (product, index) => {
-					return product.productContent.map(item => <Chip
-							key={index}
-							label={`${product.productReleaseDate}  •  ${product.productId} #${item.productPosition}  •  ${item.rarities.join(', ')}`}
-							onClick={ () => setTimeout( () => window.location.assign(`/product/${product.productId}#${cardID}`), 250 ) }
-					/>)
-					})
-			}
-
-			populateProductChips(productInfo).then(productInfoChips => setPackInfoChips(productInfoChips))
+			populateProductChips(productInfo, cardID).then(productInfoChips => setPackInfoChips(productInfoChips))
 		}
 	}, [productInfo])
 
@@ -98,11 +101,6 @@ const Card = ( { match, history } ) =>
 	useEffect( () => {
 		if ( banListInfoChips.length === 0 )
 		{
-			async function populateBanListChips(banListInfo)
-			{
-				return banListInfo.map( (item, index) => <Chip key={index} label={`${item.banListDate}  •  ${item.banStatus.charAt(0)}`} />)
-			}
-
 			populateBanListChips(banListInfo).then(banListInfoChips => setBanListInfoChips(banListInfoChips))
 		}
 	}, [banListInfo])
@@ -131,47 +129,12 @@ const Card = ( { match, history } ) =>
 					/>
 				}
 				twoThirdComponent={
-					<RightBoxPaper>
-						<RightBoxHeaderTypography variant='h4' >
-							Explore
-						</RightBoxHeaderTypography>
-						<RightBoxSubHeaderTypography variant='subtitle1' style={{marginBottom: '3rem'}} >
-							Related Content For <i>{cardName}</i>
-						</RightBoxSubHeaderTypography>
-
-						<Grid container spacing={1} >
-							<Grid item xs={12} sm={12} md={12} lg={6} xl={6}  style={ { display: 'inline-grid' } } >
-								{(isLoading)? undefined
-								: <CardInformationSection
-									isLoading={isLoading}
-									hasInfo={ (productInfo.length === 0)? false : true }
-									infoChips={productInfoChips}
-									headerText={'Products'}
-									noInfoText={'Not Found In Any Product'}
-									background='#a4508b'
-									backgroundImage='linear-gradient(326deg, #a4508b 0%, #5f0a87 74%)'
-								/>
-								}
-							</Grid>
-
-							<Grid item xs={12} sm={12} md={12} lg={6} xl={6} style={ { display: 'inline-grid' } } >
-								{(isLoading)? undefined
-								: <CardInformationSection
-									isLoading={isLoading}
-									hasInfo={ (banListInfo.length === 0)? false : true }
-									infoChips={banListInfoChips}
-									headerText={'Ban Lists'}
-									noInfoText={`No Instances of ${cardName} Being Forbidden, Limited, or Semi-Limited`}
-									background='#fc9842'
-									backgroundImage='linear-gradient(315deg, #fc9842 0%, #fe5f75 74%)'
-								/>
-								}
-
-							</Grid>
-							<Footer />
-
-						</Grid>
-					</RightBoxPaper>
+					<CardInformationRelatedContent cardName={cardName}
+						isLoading={isLoading}
+						productInfo={productInfo}
+						productInfoChips={productInfoChips}
+						banListInfo={banListInfo}
+						banListInfoChips={banListInfoChips} />
 					}
 				/>
